Only remove 99-netbox.yml if /ansible/inventory is writable (#115)

+++ files/scripts/run.sh
@@ -16,11 +16,13 @@ ANSIBLE_DIRECTORY=/ansible
 CONFIGURATION_DIRECTORY=/opt/configuration
 ENVIRONMENTS_DIRECTORY=$CONFIGURATION_DIRECTORY/environments
 
+export ANSIBLE_INVENTORY=$ANSIBLE_DIRECTORY/inventory
+
 if [[ -e /ansible/ara.env ]]; then
     source /ansible/ara.env
 fi
 
-if [[ ! -e /run/secrets/NETBOX_TOKEN ]]; then
+if [[ ! -e /run/secrets/NETBOX_TOKEN && -w $ANSIBLE_INVENTORY ]]; then
     rm -f /ansible/inventory/99-netbox.yml
 fi
 
@@ -28,8 +30,6 @@ export ANSIBLE_CONFIG=$ENVIRONMENTS_DIRECTORY/ansible.cfg
 if [[ -e $ENVIRONMENTS_DIRECTORY/$ENVIRONMENT/ansible.cfg ]]; then
     export ANSIBLE_CONFIG=$ENVIRONMENTS_DIRECTORY/$ENVIRONMENT/ansible.cfg
 fi
-
-export ANSIBLE_INVENTORY=$ANSIBLE_DIRECTORY/inventory
 
 if [[ -w $ANSIBLE_INVENTORY ]]; then
     rsync -a /ansible/group_vars/ /ansible/inventory/group_vars/
